third_party/blink/web_tests/http/tests/xmlviewer/dumpAsText/xsl-stylesheet-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
This tests that xml viewer is not used when there is an XSL stylesheet. SUCCESS
